pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbt-schemify"
version = "0.1.0"
description = "Generate and update dbt schema.yml files from a template, manifest, and database"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["dbt", "schema", "yaml", "analytics", "data engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
postgres  = ["psycopg2-binary"]
redshift  = ["psycopg2-binary"]
snowflake = ["snowflake-connector-python"]
bigquery  = ["google-cloud-bigquery"]
duckdb    = ["duckdb"]

[project.scripts]
dbt-schemify = "dbt_schemify.main:main"

[project.urls]
Repository = "https://github.com/katiaviene/dbt-schemify"
Issues     = "https://github.com/katiaviene/dbt-schemify/issues"

[tool.setuptools.packages.find]
include = ["dbt_schemify*"]
